+++ src/SchedulerTest.php
@@ -95,6 +95,8 @@ final class SchedulerTest extends TestCase
             // Issue #76, DOW and DOM do not support ?
             ['0 12 * * ?', strtotime('2020-08-20 00:00:00'), '2020-08-20 12:00:00', false],
             ['0 12 ? * *', strtotime('2020-08-20 00:00:00'), '2020-08-20 12:00:00', false],
+            // Issue #88 and #89
+            ['41-59/24 5 * * *', strtotime('2021-08-25 10:00:00'), '2021-08-26 05:41:00', false],
         ];
     }