+++ CHANGELOG.md
@@ -16,7 +16,7 @@ All Notable changes to `cron` will be documented in this file
 ### Fixed
 
 - Improve parser range validation the lower bound should always be lower or equal to the upper bound.
-- **[BC Break]** Forbids wrap around when steps are higher than the range see for reference [#88](https://github.com/dragonmantank/cron-expression/issue/88/)
+- Fix wrap around when steps are higher than the range see for reference [#88](https://github.com/dragonmantank/cron-expression/issue/88/)
 - **[BC Break]** Internal rewrite CRON Fields are now value objects.
 - **[BC Break]** `Expression` field related methods no longer returns string but `CronField` object.
 - **[BC Break]** All validators are now value object for CRON field expression (ie: `CronFieldValidator` is now `CronField`)